Extract Hours Spent field into dedicated component
We're going to be embellishing this with more content -- it feels time to component-ize
